Attempt to switch to functional model
We will have two functional models which share the same weights,
one for training and the other for sampling. The training model
has a set batch size (by default 32) and the sample model has a
batch size of 1. It is not possible to do this as a single model
when using stateful RNNs.